requirements.txt
cffi==1.17.1
colorama==0.4.6
crayons==0.4.0
cryptography==44.0.0
numpy==2.2.0
opencv-python==4.10.0.84
piexif==1.1.3
pillow==10.4.0
pycparser==2.22
stegano==0.11.4
